+++ memfile.hpp
@@ -4,13 +4,13 @@
 struct memfile {
 	int fd = 0;
 	char *map = NULL;
-	long len = 0;
-	long off = 0;
-	unsigned long tree = 0;
+	size_t len = 0;
+	size_t off = 0;
+	size_t tree = 0;
 };
 
 struct memfile *memfile_open(int fd);
 int memfile_close(struct memfile *file);
-int memfile_write(struct memfile *file, void *s, long len);
+ssize_t memfile_write(struct memfile *file, void *s, size_t len);
 
 #endif
